+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './keyBitLengthToGCMType';
+export * from './NodeAesCipher';
 export * from './NodeCbcCipher';
 export * from './NodeGcmCipher';
 export * from './nodeRrandomBytes';